+++ Tests/GRPCCodeGenTests/Internal/Translator/IDLToStructuredSwiftTranslatorSnippetBasedTests.swift
@@ -21,164 +21,6 @@ import XCTest
 @testable import GRPCCodeGen
 
 final class IDLToStructuredSwiftTranslatorSnippetBasedTests: XCTestCase {
-  func testImports() throws {
-    var dependencies = [Dependency]()
-    dependencies.append(Dependency(module: "Foo", accessLevel: .public))
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .typealias, name: "Bar"),
-        module: "Foo",
-        accessLevel: .internal
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .struct, name: "Baz"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .class, name: "Bac"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .enum, name: "Bap"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .protocol, name: "Bat"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .let, name: "Baq"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .var, name: "Bag"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .func, name: "Bak"),
-        module: "Foo",
-        accessLevel: .package
-      )
-    )
-
-    let expectedSwift =
-      """
-      /// Some really exciting license header 2023.
-
-      public import GRPCCore
-      public import Foo
-      internal import typealias Foo.Bar
-      package import struct Foo.Baz
-      package import class Foo.Bac
-      package import enum Foo.Bap
-      package import protocol Foo.Bat
-      package import let Foo.Baq
-      package import var Foo.Bag
-      package import func Foo.Bak
-
-      """
-    try self.assertIDLToStructuredSwiftTranslation(
-      codeGenerationRequest: makeCodeGenerationRequest(dependencies: dependencies),
-      expectedSwift: expectedSwift,
-      accessLevel: .public
-    )
-  }
-
-  func testPreconcurrencyImports() throws {
-    var dependencies = [Dependency]()
-    dependencies.append(
-      Dependency(
-        module: "Foo",
-        preconcurrency: .required,
-        accessLevel: .internal
-      )
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .enum, name: "Bar"),
-        module: "Foo",
-        preconcurrency: .required,
-        accessLevel: .internal
-      )
-    )
-    dependencies.append(
-      Dependency(
-        module: "Baz",
-        preconcurrency: .requiredOnOS(["Deq", "Der"]),
-        accessLevel: .internal
-      )
-    )
-    let expectedSwift =
-      """
-      /// Some really exciting license header 2023.
-
-      public import GRPCCore
-      @preconcurrency internal import Foo
-      @preconcurrency internal import enum Foo.Bar
-      #if os(Deq) || os(Der)
-      @preconcurrency internal import Baz
-      #else
-      internal import Baz
-      #endif
-
-      """
-    try self.assertIDLToStructuredSwiftTranslation(
-      codeGenerationRequest: makeCodeGenerationRequest(dependencies: dependencies),
-      expectedSwift: expectedSwift,
-      accessLevel: .public
-    )
-  }
-
-  func testSPIImports() throws {
-    var dependencies = [Dependency]()
-    dependencies.append(
-      Dependency(module: "Foo", spi: "Secret", accessLevel: .internal)
-    )
-    dependencies.append(
-      Dependency(
-        item: .init(kind: .enum, name: "Bar"),
-        module: "Foo",
-        spi: "Secret",
-        accessLevel: .internal
-      )
-    )
-
-    let expectedSwift =
-      """
-      /// Some really exciting license header 2023.
-
-      public import GRPCCore
-      @_spi(Secret) internal import Foo
-      @_spi(Secret) internal import enum Foo.Bar
-
-      """
-    try self.assertIDLToStructuredSwiftTranslation(
-      codeGenerationRequest: makeCodeGenerationRequest(dependencies: dependencies),
-      expectedSwift: expectedSwift,
-      accessLevel: .public
-    )
-  }
-
   func testGeneration() throws {
     var dependencies = [Dependency]()
     dependencies.append(
